move applyDamage to actor
